+++ server/application.production.js
@@ -1,8 +1,7 @@
-/* eslint-disable no-console */
 import express from 'express';
 import path from 'path';
+import configuration from '../utilities/configuration';
 import bodyParser from 'body-parser';
-import open from 'open';
 import logger from '../utilities/logger';
 
 import v1router from './routes/routes-v1';
@@ -11,7 +10,7 @@ const application = express();
 application.use(bodyParser.urlencoded({ extended: true }));
 application.use(bodyParser.json());
 
-const port = process.env.PORT || 8080; //eslint-disable-line no-process-env
+const port = configuration.port;
 
 application.use('/api/v1', v1router);
 application.use(express.static(path.join(__dirname, '../dist')));
@@ -22,10 +21,9 @@ application.get('*', (request, response) => {
 });
 
 application.listen(port, (error) => {
-  if (!!error) {
+  if (error) {
     logger.error(error);
   }
-  open(`http://localhost:${port}`);
   logger.info(`Serving API AT http://localhost:${port}`);
 });