+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.6.2 (March 11, 2017)
+
+- Fix webpack babel loader to only load once based for `js` and `jsx` extensions.  
+- Clean up environment variable proliferation throughout the application
+
 ## 1.6.1 (March 11, 2017)
 
 - Add Infrastructure Diagram which provides a high level overview of what the infrastructure goals are for the application.